requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-modeltranslation==0.18.2
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
